[Refacto] Use a app folder to put our code in
